Refactored search component
